Move LegacyInterceptorProvider to its own file
